+++ src/domain/job.rs
@@ -1,7 +1,7 @@
-use crate::{
-    db::entities,
-    repositories::{CreateJobParams, ListJobsParams, PatchJobParams},
-};
+use crate::db::entities;
+
+pub mod dao;
+use dao::*;
 
 #[async_trait::async_trait]
 pub trait JobRepository {
